Added stuff

+++ Slice/Models/Nodes/ValueNodes/DecimalNode.cs
@@ -2,9 +2,9 @@ using Slice.Models.Nodes;
 
 namespace Slice.Models.Nodes.ValueNodes;
 
-public class DecimalNode : Node<double>
+public class DecimalNode : Node<decimal>
 {
-    public DecimalNode(double value) : base(value)
+    public DecimalNode(decimal value) : base(value)
     { }
 
     public override void Print(string padding)
